requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
joblib
reportlab
python-dateutil
chardet